+++ main.py
@@ -10,10 +10,10 @@ from core.controller.camera import Camera
 from core.prefabs.sprites import *
 from world.chunk import Chunk
 from world.material import Material
-from world.materials import Materials
+from world.entity.materials import Materials
 from world.world import World
 from world.spawner import Spawner
-from world.entitytypes import EntityTypes
+from world.entity.entitytypes import EntityTypes
 from cfg.cfgparser import CfgParser
 
 
@@ -42,8 +42,6 @@ class Game:
 		self.load_data()
 		self.world = None
 
-	# TODO: for loop to populate assets
-
 		# Make console
 		self.console = console.Console(self)
 		self.consoleThread = threading.Thread(target=self.console.run, daemon=True)
@@ -56,22 +54,17 @@ class Game:
 
 		# Initialize config
 		self.cpc = ConfigParser()  # ConfigParserControls
+		self.cpc = ConfigParser()  # ConfigParserControls
 		self.cpc.read(path.join(path.dirname(__file__), 'cfg/controls.ini'))
 		cfgp = CfgParser(self, path.join(game_folder, 'cfg/autoexec.cfg'))
 		cfgp.read()
 
-	# self.map = tilemap.Map(path.join(game_folder, 'saves/map3.txt'))
-	# self.player_img = pygame.image.load(path.join(assets_folder, 'visual/')).convert_alpha()
-	# self.player_img = pygame.transform.scale(assets.get_asset_from_name(self.graphics, 'player1').image, (64, 64))
-
 	def new(self):
-
-
-		# initialize all variables and do all the setup for a new game
+		# Initialize all variables and do all the setup for a new game
 		self.sprites = pygame.sprite.Group()
 		self.walls = pygame.sprite.Group()
 		self.trees = pygame.sprite.Group()
-		self.world = World("test/world1")
+		self.world = World(path.join(path.dirname(__file__), "saves/world1"))
 		self.world.load()
 		self.player = Player(self, 20, 20, 0, 350, 0, 0)
 		self.spawner = Spawner(self, 64, 1)
