+++ world/gen/generator.py
@@ -2,12 +2,12 @@ from opensimplex import OpenSimplex
 
 from world.block import Block
 from world.chunk import Chunk
-from world.materials import Materials
-from world.enemy import Enemy
-from world.entitytypes import EntityTypes
+from world.entity.materials import Materials
+
+import json
+
 
 class Generator:
-
 	def __init__(self, seed):
 		self.seed = seed
 		self.noise = OpenSimplex(seed=seed)
@@ -19,7 +19,7 @@ class Generator:
 		return self.noise.noise2d(x, y)
 
 	def generateChunk(self, x: int, y: int):
-		print(f"generating chunk {x},{y}")
+		print(f"generating chunk ({x},{y})")
 		chunk: Chunk = Chunk([[Block(Materials.GRASS.value) for x in range(16)] for y in range(16)])
 		for dx in range(16):
 			for dy in range(16):
@@ -28,4 +28,13 @@ class Generator:
 					chunk.setBlock(dx, dy, Block(Materials.GRASS.value))
 				else:
 					chunk.setBlock(dx, dy, Block(Materials.WALL.value))
+		return chunk
+
+	def specialGen(self, x, y):
+		print(f"generating chunk ({x},{y})")
+		cfg = json.loads(open("/world/dungeon/prefabs/room0.json", "r").read())
+		chunk: Chunk = Chunk([[Block(Materials.GRASS.value) for x in range(16)] for y in range(16)])
+		for dx in range(16):
+			for dy in range(16):
+				chunk.setBlock(dx, dy, Block(Materials[cfg[str(dx)][dy].upper()].value))
 		return chunk
